requirements.txt
numpy==1.19.5
opencv_python_headless==4.5.5.64
pytorch_pretrained_bert==0.6.2
scipy==1.5.4
tqdm==4.64.0
cython
